requirements.txt
asgiref==3.8.1
Django>=5.2.2
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
celery[redis]
redis
djangorestframework
pandas
openpyxl    
